Save/Load framerate information
Set Piskel's framerate option when loading a spritesheet, and save Piskel's framerate option to the animation metadata on a save event.

Required adding an onComplete callback to PiskelApi.loadSpritesheet, because the save that happened mid-load was setting the incorrect framerate in the metadata.

Remaining problem: Changing the framerate in Piskel does not trigger a save event, so the external preview does not immediately update.