Polyfill ES6 for embedded blocks
Renames and repurposes 'react-only.js' bundle (which was only used by embedded blocks) to the 'embedBlocks.js' bundle, and includes the babel-polyfill in it (which is not otherwise provided on that page when loaded in an iframe).

Verified problem is solved on IE11 and Safari8Yosemite on SauceLabs, and things still work in Chrome latest on local.

pair: bjordan