Simplify cloneAnimation
Instead of doing its own call to the animation API to clone an animation, simply do a local clone of the animation, mark it 'not saved' and let the regular save system do the rest.

Later we can be more efficient by doing a copy request on S3 (which costs less than another upload) but for now this simplifies the whole model on the client.