Fix saving in non-gamelab projects
The sourceHandler _must_ call the callback passed to getAnimationMetadata one way or another.  I missed that I was failing to do this for apps that don't have animation metadata at all (I fixed this for the pixelation widget which defines its own source handler, but not for this generic case).  Now the callback is called either way, so saves don't hang.