survey: no results for < 5 students
Flip the nested iteration order so that we first loop through all students, then examine the sublevel results for each.  This way we can easily suppress results for surveys that have been submitted by fewer than five students in a section.

(I puzzled over how to do this neatly for about a week, but it turned out to be quite simple.)