Refactor: Extract business logic state queries
Move decisions based on redux state into pure functions in stateQueries.js (which happens to be a nice way to document our business logic).  Test them.

Adds a test utility for checking every permutation of a set of boolean arguments, which helps prove that these functions do what we expect.