Prevent Excess minHeight in Star Wars Top Instructions
When we detect images in Top Instructions, we apply a min-height to the
element to make the inevitable resize on image load less drastic.

However, this means we also apply that min height when we simply have
small inline images, as for Star Wars. To fix, we add an
"hasInlineImages" flag, and only apply the min height when false.