survey: teacher explanation viewing student survey.  Fix dots.
The teacher now sees an explanation that they can't see a student's survey, and nothing more, when attempting to see the student's view of the survey.

Also fix a couple issues in the dots.  Firstly, make the dots click handler go to the correct URL with the teacher's view of the student's work, rather than the teacher's view of their own work, matching the URL that was already shown correctly when hovering.  Secondly, make the click handler do nothing when the level is locked.