Add separate "Teacher View" button in lobby
Use a separate "Teacher View" button in the lobby to display the log browser with teacher-view features, instead of making them 'always on' for the shard owner.  This lets us enable teacher view in levels where the log browser is disabled, and leaves the teacher an option for having a live log browser on the projector without revealing sender names.