Make p5.allSprites use GameLabGroup customizations
In GameLab we extend p5.play's `Group` class by replacing its constructor on the p5 prototype, which allows us to give groups additional methods and custom behavior.  This works for all user-created groups; however, I discovered that the `Game.allSprites` group (actually `<p5 context>.allSprites`) does not include our modificationst to the Group class.

The reason for this was that `allSprites` is lazy-created using the closed-over original definition of `Group` instead of a reference to `p5.prototype.Group`, which we replace.  This is the only reference in p5.play to the closed-over version of Group, and it's not accessible from outside that module (so we couldn't extend it).  The fix is to construct `allSprites` using p5.prototype.Group, the same way everything else has to.

Same fix opened upstream in [molleindustria/p5.play#108](https://github.com/molleindustria/p5.play/pull/108).