Prevent setting Maze image hrefs to undefined
In some cases when we want to hide a tile image in maze, such as when
the farmer "flattens" a dirt pile and we want to remoe the pile image,
we had been setting the href of the image to `undefined`. This caused no
user-visible error as the image was simultaneously being hidden, but
still caused the page to try to load the url 'undefined' and threw a
404.

The simple fix is to instead set the href to the empty string.