Pin react-addons-shallow-compare
Even though we don't use this directly, some of our dependencies (react-color in particular) do, and don't pin it to a particular version.  As a result, you'll find get this npm weirdness:

* Apps depends on and installs React 15.3.0
* react-color depends on react-addons-shallow-compare@^15.0
* react-addons-shallow-compare@^15.0 gets react-addons-shallow-compare 15.3.1
* react-addons-shallow-compare 15.3.1 has a peerDependency to React 15.3.1
* npm explodes because its peerDependency isn't satisfied.

Apparently the npm@2 solution to this is to depend on the grandchild package at the version you'd like to pin it to.  Doing so yields this outcome:

* Apps depends on and installs React 15.3.0
* Apps depends on and installs react-addons-shallow-compare 15.3.0
* react-color depends on react-addons-shallow-compare@^15.0, which is now satisfied
* Everyone is happy