Unsubscribe from Pusher when navigating away
Our Pusher metrics suggest we are leaving many connections open.  While investigating I noticed that our async disconnect explicitly unsubscribes from Pusher but our syncrhonous disconnect does not. I don't think this is the whole problem (I would expect them to expire these connections faster) but it shouldn't hurt.

`disconnectFromShard` is called in our `onUnload` handler as the user navigates away from the page, which is the typical way for students to disconnect. Anything in this handler has to be a synchronous operation, and even then it's not really guaranteed to work.  The API for disconnecting from Pusher is synchornous in any case, (see `NetSimShard.disconnect()`) and, I think, worth trying here, if only to make the synchronous disconnect as much like the async disconnect as possible.