Try to limit real Redis to development, test envs
Not sure this will work for unit tests since I think we force the environment to be test.  However, this should be neither here nor there - even when using real Redis, these tests always attempt to connect to Redis on localhost, when in production we keep Redis on separate nodes, so it's safe.