Make move students form submit correct boolean
The second half of this bug - the `getStayEnrolledInCurrentSection` helper was branching on a condition that will virtually never be true.  It meant to check for the string "Other Teacher" (as `getNewSectionCode` does immediately above) but instead was checking against the manually entered section code, something like "XYZABC". As a result, the form _always_ sent false... which the server was also mishandling, masking this bug.  Fun!