Only calculate importable screen after fetch and show import status
This fixed a bug where the import dialog would briefly recompute what
screens could be imported after the import was finished, causing a bunch
of conflicting ID errors to show up erroneously.

I also switched to disabling all inputs when we are in the process of
importing, along with adding a spinner to the import button.