teacher dashboard: pull assessment details for any script.
It turns out that we sometimes move multi-page assessments between scripts.  The dashboard API to roll up assessment results was hitting an error when it encountered a student result in such a level.  Now, we reuse an existing retrieval of the UserLevel that doesn't check which script the level was in when submitted.