Add option to hide source code for embedded project
When a user (teacher) goes to the advanced sharing options for a
project, and clicks on the embed tab, they will now get a checkbox
to toggle whether the source code for their project should be made
available.

When source code is hidden, we modify the embed url to use an encoded
channel id and have a ?nosource query parameter. The rails endpoint
which loads the embed page looks at the query parameter and will decode
the channel id properly.

The encoding is not intended to be remotely secure in any way. It's just
meant to make it a little bit harder for students to open the embedded
app in a new page and just remove the ?nosource query parameter to see
the source. We use a simple cipher to convert ascii letters and digits.

Since there is a lot of javascript which depends on the channel id being
parsed directly out of the url (see initApp/project.js), we have to also
decode the channelId on the client side. These decoded channel ids end
up getting used to make a lot of ajax requests to various backends,
including things like firebase where we can't do server-side channel id
decoding. So it's really just not possible to secure the source code
using encoded/encrypted channel ids. To do this securely, we would need
to implement an actual ACL system. But that is far beyond the scope of
this work item and has been deemed unnecessary.