Allow shrinkable Top Instructions
When Top Instructions are in uncollapsed mode, it is possible for the
content within them to be adjusted in such a way that the current height
becomes illegal. The easiest example of this is by clicking on the
lightbulb to make the "Do you want a hint?" dialog pop up, expanding the
instructions to their new maximum possible height, and then dismissing
the dialog. You are left with instructions taller than their content.

This PR looks at those cases when we are making an update _without
changing the collapsed state_, and updates the height as allowed.