Don't hide overlay when Video Modal closed
Previously, the overlay listened for a global document click event to
close itself, and the "ok" button was purely cosmetic.

Unfortunately, this meant that when the video modal was present, closing
the modal would close both it AND the overlay underneath it.

To prevent this, the overlay now closes when a click happens on either
it or the now-actually-useful "ok" button.