Restrict accepted file types [skip ui]
Adds a filter to the file upload dialog to only allow .png, .jpg, .jpeg files to be selected (by default).  Since this is easy to work around, also adds a helpful error message when the uploaded file cannot be parsed for image metadata (we used to just hang in an "uploading..." state).

We're not supporting .gif files because we haven't fully implemented import to Piskel yet and behavior is inconsistent.