Fix mouseX/Y [skip ui]
The collision update brings with it a p5 version update - we're now on p5.js 0.5.4.  It turns out in [0.5.2](https://github.com/processing/p5.js/releases/tag/0.5.2) a [change to how `mouseX`/`mouseY` are tracked and updated](https://github.com/processing/p5.js/pull/1471) was introduced.  That's code we'd monkeypatched over to scale mouse coordinates with our resizable playspace.

This updates the monkeypatches (for both mouse and touch coordinates) to match the code in p5.js 0.5.4, but keeps our scaling modification intact.