learn2016: expandable list of tutorials in many languages
For non-en users, /learn and /learn/robotics now have a section at the top which lists tutorials in that language, with no filter interface.  /learn shows all tutorials for that language including robotics (so that we have a bigger list), while /learn/robotics shows only the robotics tutorials.  Next, there is a button to show tutorials in more languages, which opens the regular tutorial list with filter options.

The regular tutorial list no longer filters by current locale.  If it's seen by an en user, they'll see tutorials in all languages (and most tutorials do have an en version).  If it's seen by a non-en user, it will be part of the expanded set of tutorials in more languages, and so again en tutorials are okay.