Recover when user has progress on a deleted script
The levelbuilder homepage has been failing to render for some users because their user accounts have progress on a deleted script - that is, they still have a `user_script` entry but the associated `script` is gone.  This PR updates the `User::in_progress_and_deleted_scripts` method (which is only used when rendering the dashboard homepage for logged-in users) to skip any `user_script` entries for which an associated `script` cannot be found.

I am slightly concerned about generating additional queries here but since this helper method is only used once on the dashboard homepage and scripts are cached I think this is fine.