Fix React warning
Discovered that this warning, `Component's children should not be mutated` is a [known false positive](https://github.com/facebook/react/issues/7424) in React 15.3.0 when passing `NaN` into a DOM component.  It's fixed in 15.4 but I'm not happy about getting `NaN` here anyway.

Turns out we were computing an invalid transform matrix for the visualization when it's not visible.  All we have to do is early-out on that recomputation when we know we've got bad dimensions (meaning the svg isn't visible, or something is terribly wrong).