Add support for Blockly Blocks in markdown instructions
Markdown instructions will now attempt to parse any included XML as
blockly blocks.

Note that this includes a little weirdness/ugliness: because
instructions are rendered before the main blockspace is, these embedded
blockspaces will attempt to initialize themselves before our
mainBlockSpace. This is bad for some subtle reasons; the blockly
initialization pipeline makes a lot of assumptions about the
mainBlockSpace, including that it is the first one on the page to be
created. To get around this, we depend on
https://github.com/code-dot-org/blockly/pull/26 to add an event we can
listen on.