Handle names which are a subset of other names
Formerly, pairs of names such as "First Last","First" would throw errors
when "First Last" could not find it's rightmost unique character. It now
gracefully defaults back to using the full name in that case.

TODO (elijah) (in another PR): add functionality to allow for "First L"
in this case